recommendation_system_dishpatcher-main/recommendation_system_dishpatcher-main/requirement.txt
twilio
os
smtplib
time
re
random
requests
json
python-environ
